use OHLCVAttr instead of ohlcvSeries.getSeries
