+++ backtest/run_backtest_cancel_test.go
@@ -12,7 +12,7 @@ type testCancelMystrat struct{}
 
 func (m *testCancelMystrat) OnNextOHLCV(strategy Strategy, s pine.OHLCVSeries, state map[string]interface{}) error {
 
-	close := s.GetSeries(pine.OHLCPropClose)
+	close := pine.OHLCVAttr(s, pine.OHLCPropClose)
 
 	if *close.Val() == 14 {
 		entry1 := EntryOpts{
