+++ backtest/run_backtest_long_limit_order_test.go
@@ -13,7 +13,7 @@ type testLongLimitMystrat struct{}
 
 func (m *testLongLimitMystrat) OnNextOHLCV(strategy Strategy, s pine.OHLCVSeries, state map[string]interface{}) error {
 
-	close := s.GetSeries(pine.OHLCPropClose)
+	close := pine.OHLCVAttr(s, pine.OHLCPropClose)
 
 	if *close.Val() < 15 {
