+++ pine/series_change_test.go
@@ -23,7 +23,7 @@ func TestSeriesChangeNoData(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	src := series.GetSeries(OHLCPropClose)
+	src := OHLCVAttr(series, OHLCPropClose)
 
 	rsi, err := Change(src, 2)
 	if err != nil {
@@ -55,7 +55,7 @@ func TestSeriesChangeNoIteration(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	src := series.GetSeries(OHLCPropClose)
+	src := OHLCVAttr(series, OHLCPropClose)
 	rsi, err := Change(src, 1)
 	if err != nil {
 		t.Fatal(errors.Wrap(err, "error Change"))
@@ -108,7 +108,7 @@ func TestSeriesChangeSuccess(t *testing.T) {
 		series.Next()
 
 		for i, v := range testTable {
-			src := series.GetSeries(OHLCPropClose)
+			src := OHLCVAttr(series, OHLCPropClose)
 			vw, err := Change(src, v.lookback)
 			if err != nil {
 				t.Fatal(errors.Wrap(err, "error Change"))
@@ -159,7 +159,7 @@ func TestSeriesChangeNotEnoughData(t *testing.T) {
 	series.Next()
 	series.Next()
 
-	src := series.GetSeries(OHLCPropClose)
+	src := OHLCVAttr(series, OHLCPropClose)
 
 	vw, err := Change(src, 4)
 	if err != nil {
@@ -175,7 +175,7 @@ func BenchmarkChange(b *testing.B) {
 	start := time.Now()
 	data := OHLCVTestData(start, 10000, 5*60*1000)
 	series, _ := NewOHLCVSeries(data)
-	vals := series.GetSeries(OHLCPropClose)
+	vals := OHLCVAttr(series, OHLCPropClose)
 
 	for n := 0; n < b.N; n++ {
 		series.Next()
@@ -192,7 +192,7 @@ func ExampleChange() {
 			break
 		}
 
-		close := series.GetSeries(OHLCPropClose)
+		close := OHLCVAttr(series, OHLCPropClose)
 		chg, err := Change(close, 12)
 		if err != nil {
 			log.Fatal(errors.Wrap(err, "error change"))
